requirements.txt
asgiref==3.7.2
Django==5.0
django-appconf==1.0.6
django-imagekit==5.0.0
pilkit==3.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
